- bug fix: issue 8 annealing problem

+++ mxdc/widgets/misc.py
@@ -578,7 +578,7 @@ class CryojetWidget(gtk.Frame):
             self._annealed_time = 0
             self.cryojet.stop_flow()
             dur = max(0.0, (duration-0.5*1000))
-            self._restore_anneal_id = gobject.timeout_add(duration*1000, self._stop_anneal)
+            self._restore_anneal_id = gobject.timeout_add(int(duration*1000), self._stop_anneal)
             self._progress_id = gobject.timeout_add(1000, self._update_progress, duration)
             
     def _stop_anneal(self, obj=None):
